// src/modules/survey/domain/enums/survey.enums.ts
export enum SurveyType {
  MULTIPLE_CHOICE = "MULTIPLE_CHOICE",
  YES_NO = "YES_NO",
  WEIGHTED = "WEIGHTED",
  RATING = "RATING",
}

export enum SurveyStatus {
  DRAFT = "DRAFT",
  ACTIVE = "ACTIVE",
  CLOSED = "CLOSED",
  ARCHIVED = "ARCHIVED",
}

export enum VoteType {
  SINGLE = "SINGLE",
  MULTIPLE = "MULTIPLE",
  WEIGHTED = "WEIGHTED",
}